refactor(worker): align worker structure with control and tail model
