+++ .github/workflows/ci.yml
@@ -30,6 +30,9 @@ jobs:
       - name: Check AnyHarness old flat paths
         run: python3 scripts/check_anyharness_old_paths.py
 
+      - name: Check Proliferate Worker structure
+        run: python3 scripts/check_proliferate_worker_structure.py
+
       - name: Check frontend layer boundaries
         run: python3 scripts/check_frontend_boundaries.py
