+++ anyharness/crates/proliferate-worker/src/anyharness_client/sessions.rs
@@ -3,7 +3,7 @@ use reqwest::StatusCode;
 use serde_json::Value;
 use tokio::time::{sleep, Duration, Instant};
 
-use crate::commands::mapping::PlanDecision;
+use crate::control::commands::mapping::PlanDecision;
 use crate::error::WorkerError;
 
 use super::AnyHarnessClient;
