+++ anyharness/crates/proliferate-worker/src/cloud_client/mod.rs
@@ -30,8 +30,6 @@ pub struct CloudClient {
 #[serde(rename_all = "camelCase")]
 pub struct EnrollRequest {
     pub enrollment_token: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub sandbox_profile_id: Option<String>,
     pub machine_fingerprint: String,
     pub hostname: Option<String>,
     pub worker_version: Option<String>,
@@ -44,7 +42,6 @@ pub struct EnrollRequest {
 #[serde(rename_all = "camelCase")]
 pub struct EnrollResponse {
     pub target_id: String,
-    pub sandbox_profile_id: Option<String>,
     pub worker_id: String,
     pub worker_token: String,
     pub heartbeat_interval_seconds: u64,
@@ -53,8 +50,6 @@ pub struct EnrollResponse {
 #[derive(Debug, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct HeartbeatRequest {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub sandbox_profile_id: Option<String>,
     pub status: String,
     pub status_detail: Option<String>,
     pub worker_version: Option<String>,
@@ -77,7 +72,6 @@ pub struct DesiredVersions {
 #[serde(rename_all = "camelCase")]
 pub struct HeartbeatResponse {
     pub target_id: String,
-    pub sandbox_profile_id: Option<String>,
     pub worker_id: String,
     pub status: String,
     pub server_time: String,
