--- anyharness/crates/proliferate-worker/src/commands/mod.rs
@@ -1,2 +1,0 @@
-pub mod dispatcher;
-pub mod mapping;
