+++ anyharness/crates/proliferate-worker/src/identity/credentials.rs
@@ -3,7 +3,6 @@ use crate::{error::WorkerError, store::WorkerStore};
 #[derive(Debug, Clone)]
 pub struct WorkerIdentity {
     pub target_id: String,
-    pub sandbox_profile_id: Option<String>,
     pub worker_id: String,
     pub worker_token: String,
 }
