+++ anyharness/crates/proliferate-worker/src/lifecycle/mod.rs
@@ -1,0 +1,4 @@
+pub mod heartbeat;
+pub mod self_update;
+pub mod status;
+pub mod supervisor_mailbox;
