+++ anyharness/crates/proliferate-worker/src/lifecycle/supervisor_mailbox.rs
