+++ anyharness/crates/proliferate-worker/src/runtime.rs
@@ -1,31 +1,24 @@
-use tokio::time::{sleep, Duration};
+use tokio::time::sleep;
 use tracing::{info, warn};
 
 use crate::{
-    anyharness_client::{health as anyharness_health, AnyHarnessClient},
-    cloud_client::{heartbeat, inventory as cloud_inventory, CloudClient},
-    commands,
+    cloud_client::{inventory as cloud_inventory, CloudClient},
     config::WorkerConfig,
+    control,
     error::WorkerError,
-    identity::{credentials::WorkerIdentity, enrollment},
-    inventory,
+    identity::{self, credentials::WorkerIdentity},
+    inventory, lifecycle,
     process_lock::WorkerProcessLock,
     store::WorkerStore,
-    sync, updates, versions,
+    tail,
 };
-
-struct RuntimeHealth {
-    status: &'static str,
-    status_detail: Option<String>,
-    anyharness_version: Option<String>,
-}
 
 pub async fn run(config: WorkerConfig, once: bool) -> Result<(), WorkerError> {
     let _process_lock = WorkerProcessLock::acquire(&config.worker_db_path)?;
     let store = WorkerStore::open(config.worker_db_path.clone())?;
     let cloud = CloudClient::new(&config)?;
-    let identity = ensure_identity(&config, &store, &cloud).await?;
-    let health = runtime_health(&config).await;
+    let identity = identity::ensure_enrolled(&config, &store, &cloud).await?;
+    let health = lifecycle::heartbeat::runtime_health(&config).await;
     info!(
         healthy = health.status == "online",
         status = health.status,
@@ -39,7 +32,7 @@ pub async fn run(config: WorkerConfig, once: bool) -> Result<(), WorkerError> {
     if let Err(error) = upload_inventory(&cloud, &identity, &health).await {
         warn!(?error, "worker inventory upload failed");
     }
-    if let Err(error) = send_heartbeat(&config, &cloud, &identity).await {
+    if let Err(error) = lifecycle::heartbeat::send_once(&config, &cloud, &identity).await {
         warn!(?error, "worker heartbeat failed");
     }
     if once {
@@ -50,7 +43,7 @@ pub async fn run(config: WorkerConfig, once: bool) -> Result<(), WorkerError> {
     let command_identity = identity.clone();
     let command_store = store.clone();
     tokio::spawn(async move {
-        if let Err(error) = commands::dispatcher::run_loop(
+        if let Err(error) = control::r#loop::run_loop(
             command_config,
             command_cloud,
             command_identity,
@@ -61,71 +54,43 @@ pub async fn run(config: WorkerConfig, once: bool) -> Result<(), WorkerError> {
             warn!(?error, "worker command loop exited");
         }
     });
-    let sync_config = config.clone();
-    let sync_cloud = cloud.clone();
-    let sync_identity = identity.clone();
-    let sync_store = store.clone();
+    let tail_config = config.clone();
+    let tail_cloud = cloud.clone();
+    let tail_identity = identity.clone();
+    let tail_store = store.clone();
     tokio::spawn(async move {
         if let Err(error) =
-            sync::tailer::run_loop(sync_config, sync_cloud, sync_identity, sync_store).await
+            tail::r#loop::run_loop(tail_config, tail_cloud, tail_identity, tail_store).await
         {
-            warn!(?error, "worker event sync loop exited");
+            warn!(?error, "worker event tail loop exited");
         }
     });
     let revoked_jti_config = config.clone();
     let revoked_jti_cloud = cloud.clone();
     let revoked_jti_identity = identity.clone();
     tokio::spawn(async move {
-        if let Err(error) =
-            sync::revoked_jti::run_loop(revoked_jti_config, revoked_jti_cloud, revoked_jti_identity)
-                .await
+        if let Err(error) = control::reconcile::handlers::revoked_jti::run_loop(
+            revoked_jti_config,
+            revoked_jti_cloud,
+            revoked_jti_identity,
+        )
+        .await
         {
-            warn!(?error, "worker revoked-jti sync loop exited");
+            warn!(?error, "worker revoked-jti reconcile loop exited");
         }
     });
     loop {
-        sleep(Duration::from_secs(
-            config.heartbeat_interval_seconds.max(10),
-        ))
-        .await;
-        if let Err(error) = send_heartbeat(&config, &cloud, &identity).await {
+        sleep(lifecycle::heartbeat::interval(&config)).await;
+        if let Err(error) = lifecycle::heartbeat::send_once(&config, &cloud, &identity).await {
             warn!(?error, "worker heartbeat failed");
         }
     }
 }
 
-async fn ensure_identity(
-    config: &WorkerConfig,
-    store: &WorkerStore,
-    cloud: &CloudClient,
-) -> Result<WorkerIdentity, WorkerError> {
-    if let Some(identity) = WorkerIdentity::load(store)? {
-        if let Err(error) = config.clear_enrollment_token() {
-            warn!(
-                ?error,
-                "failed to clear enrollment token from worker config"
-            );
-        }
-        return Ok(identity);
-    }
-    let local_inventory = inventory::collect();
-    let request = enrollment::build_enroll_request(config, local_inventory)?;
-    let response = cloud.enroll(&request).await?;
-    let identity = enrollment::identity_from_response(response);
-    identity.save(store)?;
-    if let Err(error) = config.clear_enrollment_token() {
-        warn!(
-            ?error,
-            "failed to clear enrollment token from worker config"
-        );
-    }
-    Ok(identity)
-}
-
 async fn upload_inventory(
     cloud: &CloudClient,
     identity: &WorkerIdentity,
-    health: &RuntimeHealth,
+    health: &lifecycle::heartbeat::RuntimeHealth,
 ) -> Result<(), WorkerError> {
     let request = cloud_inventory::report(
         inventory::collect(),
@@ -135,78 +100,4 @@ async fn upload_inventory(
     cloud
         .upload_inventory(&identity.worker_token, &request)
         .await
-}
-
-async fn send_heartbeat(
-    config: &WorkerConfig,
-    cloud: &CloudClient,
-    identity: &WorkerIdentity,
-) -> Result<(), WorkerError> {
-    let health = runtime_health(config).await;
-    let anyharness_version = health.anyharness_version.clone();
-    let worker_version = versions::worker_version();
-    let supervisor_version = versions::supervisor_version_or_configured(&config.supervisor_version);
-    let request = heartbeat::report(
-        identity,
-        health.status,
-        health.status_detail.clone(),
-        worker_version.clone(),
-        anyharness_version.clone(),
-        supervisor_version.clone(),
-    );
-    let response = cloud.heartbeat(&identity.worker_token, &request).await?;
-    crate::observability::heartbeat_ack(&response);
-    let installed = updates::desired::InstalledVersions {
-        anyharness_version,
-        worker_version,
-        supervisor_version,
-    };
-    if let Err(error) = updates::desired::reconcile(
-        config,
-        cloud,
-        identity,
-        &response.desired_versions,
-        &installed,
-    )
-    .await
-    {
-        warn!(?error, "worker update reconciliation failed");
-    }
-    Ok(())
-}
-
-async fn runtime_health(config: &WorkerConfig) -> RuntimeHealth {
-    if config.anyharness_base_url.is_none() {
-        return RuntimeHealth {
-            status: "online",
-            status_detail: None,
-            anyharness_version: None,
-        };
-    }
-    match anyharness_version(config).await {
-        Some(version) => RuntimeHealth {
-            status: "online",
-            status_detail: None,
-            anyharness_version: Some(version),
-        },
-        None => RuntimeHealth {
-            status: "degraded",
-            status_detail: Some("AnyHarness health probe failed.".to_string()),
-            anyharness_version: None,
-        },
-    }
-}
-
-async fn anyharness_version(config: &WorkerConfig) -> Option<String> {
-    let base_url = config.anyharness_base_url.clone()?;
-    let client = AnyHarnessClient::new(base_url, config.anyharness_bearer_token.clone()).ok()?;
-    for attempt in 0..20 {
-        if let Some(version) = anyharness_health::version(&client).await {
-            return Some(version);
-        }
-        if attempt < 19 {
-            sleep(Duration::from_millis(500)).await;
-        }
-    }
-    None
 }
