--- anyharness/crates/proliferate-worker/src/sync/mod.rs
@@ -1,3 +1,0 @@
-pub mod backfill;
-pub mod revoked_jti;
-pub mod tailer;
