--- anyharness/crates/proliferate-worker/src/updates/mod.rs
@@ -1,3 +1,0 @@
-pub mod desired;
-pub mod status;
-pub mod supervisor;
