+++ scripts/max_lines_allowlist.txt
@@ -16,8 +16,8 @@ anyharness/crates/anyharness-lib/src/domains/mobility/service.rs 1534 existing A
 anyharness/crates/anyharness-lib/src/domains/plans/service.rs 628 existing AnyHarness oversized file
 anyharness/crates/anyharness-lib/src/domains/reviews/service/service_tests.rs 1519 existing AnyHarness oversized test file
 anyharness/crates/anyharness-lib/src/domains/runtime_config/service.rs 1057 existing AnyHarness oversized file
-anyharness/crates/proliferate-worker/src/commands/dispatcher.rs 1710 existing Proliferate worker oversized file
-anyharness/crates/proliferate-worker/src/commands/mapping.rs 953 existing Proliferate worker oversized file
+anyharness/crates/proliferate-worker/src/control/commands/executor.rs 1548 existing Proliferate worker oversized command executor
+anyharness/crates/proliferate-worker/src/control/commands/mapping.rs 953 existing Proliferate worker oversized command mapper
 anyharness/sdk-react/src/hooks/sessions.ts 653 existing SDK React oversized file
 anyharness/sdk/src/reducer/__tests__/transcript.test.ts 1880 existing SDK oversized test file
 anyharness/sdk/src/reducer/transcript.ts 1575 existing SDK oversized file
@@ -41,7 +41,7 @@ server/proliferate/server/cloud/runtime/bootstrap.py 922 server structure migrat
 server/proliferate/server/cloud/runtime/provision.py 1688 server structure migration debt for oversized domain-adjacent file
 server/proliferate/server/cloud/runtime_config/domain/resolver.py 611 server structure migration debt for oversized pure-domain file
 server/proliferate/server/cloud/runtime_config/service.py 1096 server structure migration debt for oversized service file
-server/proliferate/server/cloud/worker/service.py 1363 server structure migration debt for oversized service file
+server/proliferate/server/cloud/worker/service.py 1356 server structure migration debt for oversized service file
 server/proliferate/server/cloud/workspaces/models.py 934 server structure migration debt for oversized API models file
 server/proliferate/server/cloud/workspaces/service.py 2643 server structure migration debt for oversized service file
 server/tests/integration/schema_migration_assertions.py 879 existing server oversized test helper
@@ -51,7 +51,7 @@ server/tests/integration/test_billing_api.py 2053 existing server oversized inte
 server/tests/integration/test_cloud_agent_auth_api.py 1664 existing server oversized integration test
 server/tests/integration/test_cloud_api.py 2373 existing server oversized integration test
 server/tests/integration/test_cloud_commands_api.py 4930 existing server oversized integration test
-server/tests/integration/test_sandbox_profile_foundation.py 1170 existing server oversized integration test
+server/tests/integration/test_sandbox_profile_foundation.py 1224 existing server oversized integration test
 server/tests/integration/test_schema_migrations.py 925 existing server oversized integration test
 server/tests/integration/test_stripe_webhooks.py 1456 existing server oversized integration test
 server/tests/unit/test_cloud_mobility_service.py 806 existing server oversized unit test
